+++ src/App.js
@@ -4,7 +4,8 @@ import FetchData from './components/FetchData';
 import Tasks from './tasks/components/Tasks';
 import Layout from './components/Layout';
 import MainPage from './mainPages/components/MainPage';
-import LoginControl from './LoginPage/LoginControl';
+import LoginPage from './LoginPage/Login';
+import RegistrationPage from './RegistrationPage/Registration';
 import Task from './TaskInfo/components/TaskInf'
 import ProfileComponent from './ProfilePage/components/ProfileComponent';
 import Users from './Users/components/Users';
@@ -14,7 +15,8 @@ export default () => (
         <Route path='/main' component={MainPage} />
         <Route path='/tasks' component={Tasks} />
         <Route path='/ProfilePage/:id'component={ProfileComponent}/>
-        <Route path='/loginPage' component={LoginControl} />
+        <Route path='/loginPage' component={LoginPage} />
+        <Route path='/registrationPage' component={RegistrationPage} />
         <Route path='/taskInf/:id' component={Task}/>
         <Route path='/Users' component={Users} />
   </Layout>
