+++ src/LoginPage/actions.js
@@ -1,0 +1,25 @@
+const requestSignIn = 'REQUEST_SIGN_IN';
+const receiveSignIn = 'RECEIVE_SIGN_IN';
+
+export const actionCreators = {
+    requestLogin: (username, pass) => async (dispatch) => {
+        dispatch({ type: requestSignIn});
+
+        const url = 'https://localhost:44331/api/login';
+        const response = await fetch(url, {
+            method: 'POST',
+            mode: 'cors',
+            headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({
+                Login: username,
+                Password: pass
+            })
+        });
+        const user = await response.json();
+
+        dispatch({ type: receiveSignIn, user });
+    }
+};
