+++ src/store/configureStore.js
@@ -4,10 +4,11 @@ import { routerReducer, routerMiddleware } from 'react-router-redux';
 import * as myReducers from './reducers';
 import * as tasksReducers from '../tasks/reducers';
 import * as profileReducer from '../ProfilePage/reducers';
-import * as loginReducer from '../LoginPage/loginStore/reducers';
 import * as taskCustomerReducer from '../TaskInfo/userReducer';
 import * as usersReducers from '../Users/reducers';
-import * as commentsReducers from '../TaskInfo/commentReducer'
+import * as commentsReducers from '../TaskInfo/commentReducer';
+import * as loginReducer from '../LoginPage/reducers';
+import * as registrationReducer from '../RegistrationPage/reducers';
 
 export default function configureStore(history, initialState) {
   const reducers = {
@@ -15,10 +16,11 @@ export default function configureStore(history, initialState) {
     tasksReducers: tasksReducers.reducer,
     profilePage: profileReducer.reducer,
     loginReducer: loginReducer.reducer,
+    registrationReducer: registrationReducer.reducer,
     taskProfilePage: myReducers.reducer,
     customerOfTask: taskCustomerReducer.reducer,
     usersReducers: usersReducers.reducer,
-    commentsTask: commentsReducers.reducer
+      commentsTask: commentsReducers.reducer
   };
 
   const middleware = [
